feat(proof): dto usecase 패키지 내부로 위치 변경
