+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/createproofusecase/CreateProofUseCase.java
@@ -1,0 +1,6 @@
+package me.jinjjahalgae.domain.proof.usecase.createproofusecase;
+
+public interface CreateProofUseCase {
+
+    void execute(ProofCreateRequest request, Long contractId, Long userId);
+}
