+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/createproofusecase/ProofCreateRequest.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.dto.request;
+package me.jinjjahalgae.domain.proof.usecase.createproofusecase;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
