+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/createreproofusecase/CreateReProofUseCase.java
@@ -1,0 +1,8 @@
+package me.jinjjahalgae.domain.proof.usecase.createreproofusecase;
+
+import me.jinjjahalgae.domain.proof.usecase.createproofusecase.ProofCreateRequest;
+
+public interface CreateReProofUseCase {
+
+    void execute(ProofCreateRequest request, Long proofId, Long userId);
+}
