+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getawaitproofusecase/GetAwaitProofUseCaseImpl.java
@@ -1,11 +1,9 @@
-package me.jinjjahalgae.domain.proof.usecase.impls;
+package me.jinjjahalgae.domain.proof.usecase.getawaitproofusecase;
 
 import lombok.RequiredArgsConstructor;
-import me.jinjjahalgae.domain.proof.dto.response.ProofAwaitResponse;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.mapper.ProofMapper;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
-import me.jinjjahalgae.domain.proof.usecase.interfaces.GetAwaitProofUseCase;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
