+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getcontractorprooflistusecase/ContractorProofListResponse.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.dto.response;
+package me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase;
 
 import io.swagger.v3.oas.annotations.media.Schema;
