+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getcontractorprooflistusecase/GetContractorProofListUseCaseImpl.java
@@ -1,11 +1,9 @@
-package me.jinjjahalgae.domain.proof.usecase.impls;
+package me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase;
 
 import lombok.RequiredArgsConstructor;
-import me.jinjjahalgae.domain.proof.dto.response.ContractorProofListResponse;
 import me.jinjjahalgae.domain.proof.entities.Proof;
 import me.jinjjahalgae.domain.proof.mapper.ProofMapper;
 import me.jinjjahalgae.domain.proof.repository.ProofRepository;
-import me.jinjjahalgae.domain.proof.usecase.interfaces.GetContractorProofListUseCase;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
