+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getproofdetailusecase/GetProofDetailUseCase.java
@@ -1,0 +1,6 @@
+package me.jinjjahalgae.domain.proof.usecase.getproofdetailusecase;
+
+public interface GetProofDetailUseCase {
+
+    ProofDetailResponse execute(Long proofId, Long userId);
+}
