+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getrecentproofusecase/ProofRecentResponse.java
@@ -1,4 +1,4 @@
-package me.jinjjahalgae.domain.proof.dto.response;
+package me.jinjjahalgae.domain.proof.usecase.getrecentproofusecase;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import me.jinjjahalgae.domain.proof.enums.ProofStatus;
