+++ src/main/java/me/jinjjahalgae/domain/proof/usecase/getsupervisorprooflistusecase/SupervisorProofListResponse.java
@@ -1,7 +1,8 @@
-package me.jinjjahalgae.domain.proof.dto.response;
+package me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflistusecase;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import me.jinjjahalgae.domain.feedback.enums.FeedbackStatus;
+import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase.ProofSimpleResponse;
 
 /**
  * 감독자 달력 표시를 위한 response
