--- src/main/java/me/jinjjahalgae/domain/proof/usecase/interfaces/CreateProofUseCase.java
@@ -1,8 +1,0 @@
-package me.jinjjahalgae.domain.proof.usecase.interfaces;
-
-import me.jinjjahalgae.domain.proof.dto.request.ProofCreateRequest;
-
-public interface CreateProofUseCase {
-
-    void execute(ProofCreateRequest request, Long contractId, Long userId);
-}
