+++ src/main/java/me/jinjjahalgae/presentation/api/ProofController.java
@@ -2,9 +2,19 @@ package me.jinjjahalgae.presentation.api;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import me.jinjjahalgae.domain.proof.dto.request.ProofCreateRequest;
-import me.jinjjahalgae.domain.proof.dto.response.*;
-import me.jinjjahalgae.domain.proof.usecase.interfaces.*;
+import me.jinjjahalgae.domain.proof.usecase.createproofusecase.ProofCreateRequest;
+import me.jinjjahalgae.domain.proof.usecase.createproofusecase.CreateProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.createreproofusecase.CreateReProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getawaitproofusecase.GetAwaitProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getawaitproofusecase.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase.GetContractorProofListUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getproofdetailusecase.GetProofDetailUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getproofdetailusecase.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.getrecentproofusecase.GetRecentProofUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getrecentproofusecase.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflistusecase.GetSupervisorProofListUseCase;
+import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflistusecase.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import me.jinjjahalgae.presentation.api.docs.proof.ProofControllerDocs;
