+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/GetProofDetailSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.proof;
 
-import me.jinjjahalgae.domain.proof.dto.response.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.getproofdetailusecase.ProofDetailResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class GetProofDetailSwaggerResponse extends CommonResponse<ProofDetailResponse> {
