+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/GetSupervisorProofListSwaggerResponse.java
@@ -1,6 +1,6 @@
 package me.jinjjahalgae.presentation.api.docs.proof;
 
-import me.jinjjahalgae.domain.proof.dto.response.SupervisorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflistusecase.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 
 public class GetSupervisorProofListSwaggerResponse extends CommonResponse<SupervisorProofListResponse> {
