+++ src/main/java/me/jinjjahalgae/presentation/api/docs/proof/ProofControllerDocs.java
@@ -8,8 +8,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import me.jinjjahalgae.domain.proof.dto.request.ProofCreateRequest;
-import me.jinjjahalgae.domain.proof.dto.response.*;
+import me.jinjjahalgae.domain.proof.usecase.createproofusecase.ProofCreateRequest;
+import me.jinjjahalgae.domain.proof.usecase.getawaitproofusecase.ProofAwaitResponse;
+import me.jinjjahalgae.domain.proof.usecase.getcontractorprooflistusecase.ContractorProofListResponse;
+import me.jinjjahalgae.domain.proof.usecase.getproofdetailusecase.ProofDetailResponse;
+import me.jinjjahalgae.domain.proof.usecase.getrecentproofusecase.ProofRecentResponse;
+import me.jinjjahalgae.domain.proof.usecase.getsupervisorprooflistusecase.SupervisorProofListResponse;
 import me.jinjjahalgae.global.common.CommonResponse;
 import me.jinjjahalgae.global.security.jwt.CustomJwtPrincipal;
 import org.springframework.web.ErrorResponse;
